requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cmdstanpy==1.2.5
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.58.1
holidays==0.74
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prophet==1.1.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stanio==0.5.1
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
